pyproject.toml
[tool.poetry]
name = "core-engine"
version = "0.1.0"
description = "Core Engine for Snap Report"
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pymupdf = "^1.23.0"
numpy = "^1.26.0"
scikit-learn = "^1.4.0"
pillow = "^10.2.0"
requests = "^2.31.0"
httpx = "^0.27.0"
python-multipart = "^0.0.9"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
sqlalchemy = "^2.0.0"
aiosqlite = "^0.20.0"
google-cloud-storage = "^2.14.0"
piexif = "^1.1.3"
pyproj = "^3.6.0"
aiofiles = "^23.2.0"
aioshutil = "^1.3"
prometheus-client = "^0.23.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
